requirements.txt
Django==3.2.11
psycopg2==2.7.7
psycopg2-binary==2.9.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
drf-nested-routers==0.93.4
django-filter==21.1